+++ app/helpers/nfg_ui/application_helper.rb
@@ -21,7 +21,6 @@ module NfgUi
       else
         ApplicationController.new.view_context
       end
-      # @ui ||= NfgUi::UI::Base.new(controller.view_context)
     end
   end
 end